Create pull_weather_data.py
working version of pull_weather_data, currently scrapes all weather data on the NCEI global summary of the day daily product for Colorado springs municipal airport and saves it to an sqlite database. This will be bulk storage of raw data. the daily product stores data by the year for ALL stations, so this script checks each year for the desired station ID, pulls the CSV data stored for it and compiles all of it into one location.