Create url_scrape.py
scrapes https://www.ncei.noaa.gov/data/global-summary-of-the-day/access/ and saves a URL for the CSV that stores the daily weather data for the year. scraping for the URLs takes a long time, so this script will save the list as a json file so it only needs to scrape the website once.